+++ lib/workling/remote.rb
@@ -10,6 +10,7 @@ module Workling
     mattr_accessor :dispatcher
     @@dispatcher = Workling.default_runner
     
+    # generates a unique identifier for this particular job. 
     def self.generate_uid(clazz, method)
       uid = ::Digest::MD5.hexdigest("#{ clazz }:#{ method }:#{ rand(1 << 64) }:#{ Time.now }")
       "#{ clazz.to_s.tableize }/#{ method }/#{ uid }".split("/").join(":")
@@ -20,7 +21,7 @@ module Workling
     def self.run(clazz, method, options = {})
       uid = Workling::Remote.generate_uid(clazz, method)
       options[:uid] = uid if options.kind_of?(Hash) && !options[:uid]
-      dispatcher.send :run_with_error_handling, clazz, method, options
+      dispatcher.run(clazz, method, options)
       uid
     end
   end