+++ script/bj_invoker.rb
@@ -4,5 +4,8 @@ message, command, args = *unnormalized.match(/(^[^ ]*) (.*)/)
 options = Hash.from_xml(args)["hash"]
 
 if workling = @routing[command]
-  workling.send @routing.method_name(command), options.symbolize_keys
+  options = options.symbolize_keys
+  method_name = @routing.method_name(command)
+
+  workling.dispatch_to_worker_method(method, options)
 end